integrated method pigiffyDigitto the class PigNumberClass
