+++ src/main/java/piglatin/PigNumberClass.java
@@ -1,28 +1,46 @@
 package piglatin;
-
-import java.io.Console;
-import java.util.Arrays;
-import java.util.StringTokenizer;
-import java.io.IOException;
-import java.util.Scanner;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.text.DecimalFormat;
+import piglatin.PigLatin;
 
 public class PigNumberClass {
 
-	int pigs = 3;
+	
+    
+    public static String pigiffyDigit(int number) {
 
+    	String word;
 
-    public static void main(String[] args) {
-    	// reading from console
-        System.out.println("I am a pig.");
-                
-    }
+        // @Shahbaz, here I call a method. From Rui
+        // pigNum.oink();
 
-    public static void oink(){
+    	switch (number) {
 
-        System.out.println("Oink oink!");
+    		case 0:  word = "zero";
+                     break;
+            case 1:  word = "one";
+                     break;
+            case 2:  word = "two";
+                     break;
+            case 3:  word = "three";
+                     break;
+            case 4:  word = "four";
+                     break;
+            case 5:  word = "five";
+                     break;
+            case 6:  word = "six";
+                     break;
+            case 7:  word = "seven";
+                     break;
+            case 8:  word = "eight";
+                     break;
+            case 9:  word = "nine";
+                     break;
+            default: word = "";
+                     break;
+        }
+
+    	String pigiffied_word = PigLatin.pigiffy(word);
+
+    	return pigiffied_word;
 
     }
